Implement Recent Activities Section and Refactor Edit Cab Dialog
- Added `RecentActivitiesSection.vue` to display the last 5 activities with formatted timestamps.
- Refactored `EditCabDialog.vue` to simplify image validation logic and improve state management when updating cab data.
- Updated `DashboardPage.vue` to integrate the new `RecentActivitiesSection` component.
