+++ Frontend/src/components/dashboard/RecentActivitiesSection.vue
